Manually update to avoid conflict errors on report_only
The results report is generated in another thread (through the
browser), so the status of client and sample objects in the
current thread is out-of-sync once the report is generated.

This commit forces the update of the objects from db so their
status is in-sync and we can perform the publish transition
manually.